delete renderQueue from base class
