+++ cocos/core/pipeline/ui/ui-stage.ts
@@ -2,12 +2,17 @@
  * @category pipeline
  */
 
+import { ccclass, property } from '../../data/class-decorator';
 import { IGFXColor, IGFXRect } from '../../gfx/define';
-import { IRenderStageInfo, RenderQueueSortMode, RenderStage } from '../render-stage';
+import { IRenderStageInfo, RenderStage } from '../render-stage';
 import { RenderView } from '../render-view';
 import { ForwardStagePriority } from '../forward/enum';
 import { UIFlow } from './ui-flow';
 import { ForwardPipeline } from '../forward/forward-pipeline';
+import { RenderQueueDesc, RenderQueueSortMode } from '../pipeline-serialization';
+import { getPhaseID } from '../pass-phase';
+import { opaqueCompareFn, RenderQueue, transparentCompareFn } from '../render-queue';
+import { IRenderPass } from '../define';
 
 const colors: IGFXColor[] = [];
 
@@ -15,21 +20,57 @@ const colors: IGFXColor[] = [];
  * @en The UI render stage
  * @zh UI渲阶段。
  */
+@ccclass('UIStage')
 export class UIStage extends RenderStage {
     public static initInfo: IRenderStageInfo = {
         name: 'UIStage',
         priority: ForwardStagePriority.UI,
-        renderQueues: [{
-            isTransparent: true,
-            stages: ['default'],
-            sortMode: RenderQueueSortMode.BACK_TO_FRONT,
-        }],
     };
+
+    @property({
+        type: [RenderQueueDesc],
+        displayOrder: 2,
+        visible: true,
+    })
+    protected renderQueues: RenderQueueDesc[] = [];
+    protected _renderQueues: RenderQueue[] = [];
 
     private _renderArea: IGFXRect = { x: 0, y: 0, width: 0, height: 0 };
 
+    public initialize (info: IRenderStageInfo) {
+        super.initialize(info);
+        this.renderQueues = [{
+            isTransparent: true,
+            stages: ['default'],
+            sortMode: RenderQueueSortMode.BACK_TO_FRONT,
+        }]
+
+        return true;
+    }
+
     public activate (pipeline: ForwardPipeline, flow: UIFlow) {
         super.activate(pipeline, flow);
+        for (let i = 0; i < this.renderQueues.length; i++) {
+            let phase = 0;
+            for (let j = 0; j < this.renderQueues[i].stages.length; j++) {
+                phase |= getPhaseID(this.renderQueues[i].stages[j]);
+            }
+            let sortFunc: (a: IRenderPass, b: IRenderPass) => number = opaqueCompareFn;
+            switch (this.renderQueues[i].sortMode) {
+                case RenderQueueSortMode.BACK_TO_FRONT:
+                    sortFunc = transparentCompareFn;
+                    break;
+                case RenderQueueSortMode.FRONT_TO_BACK:
+                    sortFunc = opaqueCompareFn;
+                    break;
+            }
+
+            this._renderQueues[i] = new RenderQueue({
+                isTransparent: this.renderQueues[i].isTransparent,
+                phases: phase,
+                sortFunc,
+            });
+        }
     }
 
     public destroy () {
